fetch/metadata/generated/form-submission-sec-fetch-mode.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Mode</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(method, url) {
    // Query parameters must be expressed as form values so that they are sent
    // with the submission of forms whose method is POST.
    const windowName = String(Math.random());
    const form = document.createElement('form');
    const submit = document.createElement('input');
    submit.setAttribute('type', 'submit');
    form.appendChild(submit);
    const win = open('about:blank', windowName);
    form.setAttribute('method', method);
    form.setAttribute('action', url);
    form.setAttribute('target', windowName);
    document.body.appendChild(form);

    Array.from(new URL(url, location.origin).searchParams)
      .forEach(([name, value]) => {
        const input = document.createElement('input');
        input.setAttribute('type', 'hidden');
        input.setAttribute('name', name);
        input.setAttribute('value', value);
        form.appendChild(input);
      });

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });

        submit.click();
      })
      .then(() => {
        form.remove();
        win.close();
      });
  }
  const responseParams = {
    mime: 'text/html',
    body: `<script>opener.postMessage('done', '*')</${''}script>`
  };

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    return induceRequest('GET', url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'GET');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, [], responseParams);
    return induceRequest('POST', url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'POST');
  </script>
  </body>
</html>
